Add buildroot as a submodule and restructure repo
To make this project easily buildable I added a submodule for the
buildroot dependency and then restructured all of the code so that it's
in the recommended format for a buildroot external tree. I also added
configure, build, and run scripts that should make it easier to set up
the project and use it.