Clean up build files and lay out makefiles
Fix some issues with build scripts, such as an unnecessary rm in the run
script and a missing existence check in the post-build script.

Fix a type in the name of the external buildroot tree description file.

Create some placeholder makefiles for building the bomb package that
will be used for the actual game.